+++ src/api/fetch/auth/api/useApiKakaoLogin.ts
@@ -1,10 +1,5 @@
-import useAppMutation from "@/api/_base/query/useAppMutation";
-import { ApiBaseResponseType } from "@/api/_base/types/ApiBaseResponseType";
-import { useToast } from "@/context/ToastContext";
-import { useRouter } from "next/navigation";
-import { AUTH_LOGIN_SUCCESS_EVENT } from "@/constants";
 import { KakaoLoginResponseType } from "../types/KakaoLoginResponseType";
-import { AxiosError } from "axios";
+import { useApiSocialLogin } from "./useApiSocialLogin";
 
 interface KakaoRequestType {
   code: string;
@@ -15,30 +10,5 @@ interface KakaoRequestType {
   marketingConsent?: boolean;
 }
 
-export const useApiKakaoLogin = () => {
-  const router = useRouter();
-
-  const { addToast } = useToast();
-
-  return useAppMutation<
-    KakaoRequestType,
-    KakaoLoginResponseType,
-    AxiosError<ApiBaseResponseType<null>>
-  >("auth", "/auth/kakao", "post", {
-    onSuccess: () => {
-      if (typeof window !== "undefined") {
-        window.dispatchEvent(new CustomEvent(AUTH_LOGIN_SUCCESS_EVENT));
-      }
-    },
-    onError: (error) => {
-      if (error.response?.data.code === "AUTH400-KAKAO_CODE_INVALID")
-        addToast("카카오 인증코드가 유효하지 않거나 이미 사용되었어요", "warning");
-      else if (error.response?.data.code === "AUTH500-KAKAO_USERINFO_FAILED")
-        addToast("카카오 사용자 정보 조회에 실패했어요", "warning");
-      else if (error.response?.data.code === "AUTH409-EMAIL_RECENTLY_DELETED")
-        addToast("탈퇴한 사용자에요", "warning");
-      else addToast("잠시 후 다시 시도해 주세요", "warning");
-      router.replace("/login");
-    },
-  });
-};
+export const useApiKakaoLogin = () =>
+  useApiSocialLogin<KakaoRequestType, KakaoLoginResponseType>("kakao");
